feat(universal-modal): mobile file renaming
