+++ packages/universal-modal/src/mobile/components/footer/index.module.css
